Updated UTM Implementation
