+++ conversions.h
@@ -1,0 +1,12 @@
+//
+
+#ifndef CONVERSIONS_H
+#define CONVERSIONS_H
+
+#include <math.h>
+
+inline double DegreesToRadians(double Degrees) {
+    return Degrees * ((double)(M_PI) / 180);
+};
+
+#endif
